+++ stag_lib/cluster.h
@@ -61,6 +61,35 @@ namespace stag {
    */
   std::vector<stag_int> spectral_cluster(stag::Graph* graph, stag_int k);
 
+  /**
+   * Find the Cheeger cut in a graph.
+   *
+   * Let \f$G = (V, E)\f$ be a graph and \f$\mathcal{L}\f$ be its normalised Laplacian
+   * matrix with eigenvalues \f$0 = \lambda_1 \leq \lambda_2 \leq \ldots \leq \lambda_n\f$.
+   * Then, Cheeger's inequality states that
+   *
+   * \f[
+   *   \frac{\lambda_2}{2} \leq \Phi_G \leq \sqrt{2 \lambda_2},
+   * \f]
+   *
+   * where
+   *
+   * \f[
+   *    \Phi_G = \min_{S \subset V} \phi(S)
+   * \f]
+   *
+   * is the conductance of \f$G\f$. The proof of Cheeger's inequality is
+   * constructive: by computing the eigenvector corresponding to \f$\lambda_2\f$,
+   * and performing the sweep set operation, we are able to find a set \f$S\f$
+   * with conductance close to the optimal. The partition returned by this
+   * algorithm is called the 'Cheeger cut' of the graph.
+   *
+   * @param graph the graph object to be partitioned
+   * @return A vector giving the cluster membership for each vertex in the graph.
+   *         Each entry in the vector is either \f$0\f$ or \f$1\f$ to indicate
+   *         which side of the cut the vertex belongs to.
+   */
+  std::vector<stag_int> cheeger_cut(stag::Graph* graph);
 
   /**
    * Local clustering algorithm based on personalised Pagerank.
@@ -158,21 +187,29 @@ namespace stag {
    *
    * where \f$\phi(S)\f$ is the conductance of \f$S\f$.
    *
-   * This method is expected to be run on vectors whose support is much less
-   * than the total size of the graph. If the total volume of the support of vec
-   * is larger than half of the volume of an entire graph, then this method may
-   * return unexpected results.
+   * When the provided graph is a stag::LocalGraph, the volume of the support of
+   * the provided vector should be less than half the total volume of the graph.
+   * The method does not (and cannot) check this condition.
+   *
+   * When the provided graph is a stag::Graph, there is no restriction on the
+   * volume of the support of the provided vector.
    *
    * Note that the caller is responsible for any required normalisation of the
    * input vector. In particular, this method does not normalise the vector by
    * the node degrees.
    *
-   * @param graph a stag::LocalGraph object
+   * @param graph a stag::LocalGraph or stag::Graph object
    * @param vec the vector to sweep over
    * @return a vector containing the indices of vec which give the minimum
    *         conductance in the given graph
    */
   std::vector<stag_int> sweep_set_conductance(stag::LocalGraph* graph,
+                                              SprsMat& vec);
+
+  /**
+   * @overload
+   */
+  std::vector<stag_int> sweep_set_conductance(stag::Graph* graph,
                                               SprsMat& vec);
 
   /**