+++ stag_lib/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Issue #94](https://github.com/staglibrary/stag/issues/94): Add a subgraph method to stag Graph class
 - [Issue #93](https://github.com/staglibrary/stag/issues/93): Add methods to compute mutual information between clusters
 - [Issue #98](https://github.com/staglibrary/stag/issues/98): Add method to construct the graph union
+- [Issue #22](https://github.com/staglibrary/stag/issues/22): Add Cheeger cut method
 
 ### Fixed
 - [Issue #87](https://github.com/staglibrary/stag/issues/87): Occasional bug with sorting edgelist file